requirements.txt
Flask==2.3.3
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
faiss-cpu==1.7.4
ftfy==6.1.1
regex==2023.6.3
tqdm==4.65.0
clip-by-openai
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==20.1.0